Refactor: Overlay component detects a close event and remove quick view service
Overlay component did not detect its close. So when it close it have to 
close the view child itself. It should not have this behave.
the quick view open service used this overlay component. But it's 
responsability of product catalog which knows how to open the quick view 
by a listening to its events